chore: move window resize listener to parent tab and pass isMobile props to child instead
