+++ src/components/FilterControl.tsx
@@ -1,31 +1,13 @@
 import * as fabric from "fabric"
 import type { RootStateType } from "@/redux/store"
-import type { FilterControlType, FilterListType, LowercaseFilterIdType } from "@/types"
+import type { FilterControlTypeProps, FilterListType, LowercaseFilterIdType } from "@/types"
 import { useAppSelector, useAppDispatch } from "@/redux/hooks"
 import { setImageFilterValue } from "@/redux/selectedImageSlice"
 import toast from "react-hot-toast"
-import { useState, useEffect } from "react"
 
-export default function FilterControl(props: FilterControlType) {
-  const { id, min, max, step, newFilter } = props
+export default function FilterControl(props: FilterControlTypeProps) {
+  const { id, min, max, step, newFilter, isMobile } = props
   const filterTypeLower = id.toLowerCase()
-
-  const [isMobile, setIsMobile] = useState(false)
-
-  useEffect(() => {
-    const handleResize = () => {
-      const isMobileView = window.matchMedia('(max-width: 640px)').matches
-      setIsMobile(isMobileView)
-    }
-
-    handleResize()
-
-    window.addEventListener('resize', handleResize)
-
-    return () => {
-      window.removeEventListener('resize', handleResize)
-    }
-  }, [])
 
   // Get necessary data from Redux store
   const dispatch = useAppDispatch()
