+++ src/components/TabFilters.tsx
@@ -1,10 +1,28 @@
 import type { FilterControlType } from "@/types"
+import { useState, useEffect } from "react"
 import { filters } from "@/constants/filters"
 import FilterControl from "./FilterControl"
 
 export default function TabFilters() {
+  const [isMobile, setIsMobile] = useState(false)
+
+  useEffect(() => {
+    const handleResize = () => {
+      const isMobileView = window.matchMedia('(max-width: 640px)').matches
+      setIsMobile(isMobileView)
+    }
+
+    handleResize()
+
+    window.addEventListener('resize', handleResize)
+
+    return () => {
+      window.removeEventListener('resize', handleResize)
+    }
+  }, [])
+
   return (
-    <div className="w-full px-2">
+    <div className={`${isMobile ? "flex flex-nowrap" : "w-full px-2"}`}>
       {filters.map((filter: FilterControlType, i: number) => {
         return (
           <FilterControl
@@ -14,6 +32,7 @@ export default function TabFilters() {
             max={filter.max}
             step={filter.step}
             newFilter={filter.newFilter}
+            isMobile={isMobile}
           />
         )
       })}
